[NO-ISSUE] 환전 API 타입 및 필드 이름 변경

+++ data/exchange/src/main/java/com/ssafy/exchange/api/request/ExchangeRequest.kt
@@ -1,7 +1,7 @@
 package com.ssafy.exchange.api.request
 
 data class ExchangeRequest(
-    val transactionBalance: Int,
+    val transactionBalance: Long,
     val pinNumber: String,
 )
