+++ data/exchange/src/main/java/com/ssafy/exchange/domain/ExchangeRepository.kt
@@ -3,8 +3,8 @@ package com.ssafy.exchange.domain
 import com.ssafy.exchange.domain.model.Exchange
 
 interface ExchangeRepository {
-    suspend fun exchange(transactionBalance: Int, pinNumber: String): Result<Exchange>
-    suspend fun exchangeForeign(transactionBalance: Int, pinNumber: String): Result<Exchange>
+    suspend fun exchange(transactionBalance: Long, pinNumber: String): Result<Exchange>
+    suspend fun exchangeForeign(transactionBalance: Long, pinNumber: String): Result<Exchange>
     suspend fun getAiPrediction(): Result<String>
     suspend fun getHistoricalAnalysis(): Result<String>
 }
